Rename to duration_extension
