+++ pyproject.toml
@@ -367,7 +367,7 @@ allow-dict-calls-with-keyword-arguments = true
 
 [tool.ruff.lint.per-file-ignores]
 '__init__.py' = ['PLC0414']
-'doc/source/duration.py' = ['D102']
+'doc/source/duration_extension.py' = ['D102']
 'doc/source/make_tables.py' = ['D101', 'D102']
 'examples/*' = ['D102', 'D103', 'D205', 'D212', 'D400', 'D415', 'SIM115']
 'examples/**' = [
